requirements.txt
pytesseract==0.3.10
Pillow==9.4.0
openai==0.27.0
numpy==1.23.5
scikit-learn==1.2.2
python-dotenv==1.0.0
Flask==2.3.2
PyPDF2==3.0.1
python-docx==0.8.11
sqlite3==3.36.0  # Built-in with Python, included for clarity
beautifulsoup4==4.12.2
